+++ docs/source/changes.rst
@@ -25,6 +25,7 @@ Release 0.5.2 (under development)
   CUSOLVER functions.
 * Skip CULA-dependent unit tests when CULA isn't present.
 * CUSOLVER support for symmetric eigenvalue decomposition (enh. by Bryant Menn).
+* CUSOLVER support for matrix inversion (#198).
 
 Release 0.5.1 - (October 30, 2015)
 ----------------------------------